refactor(app): relocate screens into features/cards + enforce route-file rule (Story 16.9)
Pure refactor moving fat screens out of the app/ routing layer into features/,
plus the documented-but-missing route-file lint rule that let the debt accrue.

- move HomeScreen/CardDetail/CardEdit/Barcode screens app/ -> features/cards/screens/
  (verbatim; git-tracked renames); routes collapse to thin re-exports (AC1, AC2)
- add // Screens block to the features/cards barrel; moved screens deep-import
  siblings to avoid a barrel<->screen cycle
- add cards -> auth boundary exception (AD-2) and the route-file
  no-restricted-imports rule (AD-3, corrected flat-config scoping) to eslint (AC3, AC4)
- refactor the live app/scan.tsx shim to a hook-free <Redirect> (AD-4, AC5)
- remove app/__tests__/; co-locate tests in features/*; per ifero, app/ holds no
  test files -> app-subject tests (scan + root-layout) live in top-level test/,
  importing via @/app/... (supersedes AD-5) (AC6, AC7)
- add 3 screen test suites + extend HomeScreen; full suite 1675 green, coverage
  81.49% branch (AC8)
