+++ docs/sprint-artifacts/sprint-status.yaml
@@ -399,7 +399,7 @@ development_status:
   16-6-stand-up-penpot-design-space: backlog # KEPT IN BACKLOG for Sprint 16 (ifero, 2026-06-11) — trigger-gated: open no-seat-limit Penpot only when real visual designers want in. Draft file exists; not committed to Sprint 16.
   16-7-fix-android-beta-upload-versioncode: done # 2026-06-13: impl complete on feature/16-7-fix-android-beta-upload-versioncode — env-driven versionCode via app.config.ts (verified into merged manifest), workflow Compute steps, Fastfile signing-env guard, cicd.md track fix. dev code-review (sonnet, 3 rounds → 0) + QA (sonnet, 2 rounds → 0) APPROVED; 1535 tests green. Awaiting stakeholder approval to commit/PR. [2026-06-11 refine] CI tech-debt — Android RC alpha upload fails "Version code 1 has already been used" (prebuild regenerates build.gradle versionCode=1). Refined 2026-06-11 w/ ifero: versionCode from GITHUB_RUN_NUMBER plumbed via new app.config.ts (prod offset band to avoid cross-workflow collision); release signingConfig fix kept in scope; also fix prod lane + cicd.md doc. Diagnosed from CI run 27338428755.
   16-8-fix-cloud-sync-cold-open-race: done # 2026-06-13: latch-on-success + network gate + retryWithBackoff on feature/16-8-fix-cloud-sync-cold-open-race. Code review + QA review approved by sonnet subagents. Awaiting stakeholder approval. — bug (filed as tech debt per ifero) — cloud sync fails on every cold open, recovers on manual pull. Confirmed w/ ifero (signed-in; pull-to-refresh works; banner "Cloud sync failed. Pull to retry.") → autoTriggeredRef latch race is root cause (useCloudSync.ts:108-121). Fix: latch-on-success (root) + network-ready gate + retryWithBackoff (defense-in-depth). Refined → ready-for-dev 2026-06-11.
-  16-9-relocate-app-screens-to-features: ready-for-dev # 2026-06-18: relocate fat screens out of app/ routing layer into features/cards/screens/ (index/card[id]/card[id]edit/barcode[id]) + implement the documented-but-MISSING route-file lint rule (root cause) + remove app/__tests__ (co-locate). Drafted by Bob (SM); REFINED → ready-for-dev by Winston (Architect) same day. AD-6: sequence BEFORE 16-2 (both touch console.* in the 3 card screens). AD-2 (cards→auth exception for HomeScreen's guest/migration banners) RATIFIED by ifero 2026-06-18. Pure refactor, engine-agnostic.
+  16-9-relocate-app-screens-to-features: review # 2026-07-10 dev complete (Amelia): 4 screens → features/cards/screens/ (verbatim move), routes now thin re-exports, cards→auth boundary exception + route-file no-restricted-imports rule added to eslint, app/scan.tsx → hook-free <Redirect> (kept — live from CatalogueGrid), app/__tests__ removed (welcome dup deleted), +3 screen test suites + extended HomeScreen to restore coverage (81.49% branch). Per ifero: app/ holds NO test files — the 4 app-subject tests (scan + 3 root-layout) live in top-level test/ (import via @/app/...), superseding AD-5. lint+typecheck+1675 tests green. code-review + QA both APPROVED zero-comments. Awaiting ifero final review before PR. | 2026-06-18: relocate fat screens out of app/ routing layer into features/cards/screens/ (index/card[id]/card[id]edit/barcode[id]) + implement the documented-but-MISSING route-file lint rule (root cause) + remove app/__tests__ (co-locate). Drafted by Bob (SM); REFINED → ready-for-dev by Winston (Architect) same day. AD-6: sequence BEFORE 16-2 (both touch console.* in the 3 card screens). AD-2 (cards→auth exception for HomeScreen's guest/migration banners) RATIFIED by ifero 2026-06-18. Pure refactor, engine-agnostic.
   16-10-fix-offline-cold-start-hang: done # 2026-07-10 impl complete (Amelia) — AD-16-10-01 via useBootAuthGate (storage probe + reactive onAuthStateChange + safety timeout) + withTimeout on the update check; code-review + QA both APPROVED zero-comments; awaiting ifero review before PR. 2026-07-09 HOTFIX (CRITICAL — breaks offline-first): offline cold start hangs on the loading spinner forever. Root cause app/_layout.tsx:311 — unguarded await auth.getSession() (autoRefreshToken:true -> network token refresh, no timeout/abort); offline it never settles so setIsReady(true) (:317) never fires. REFINED -> ready-for-dev by Winston (Architect) 2026-07-09: AD-16-10-01 locks the fix — gate boot on the synchronous INITIAL_SESSION (useAuthState.ts:37-44, storage-only, no network) + safety timeout; REJECTED the getSession-timeout option (bounds but doesn't remove the hang). Preserves the no-flash welcome-gate intent (app/_layout.tsx:306-309); also fixes latent isAuthenticated staleness. Drafted by PM (John); FAST-TRACKED as a hotfix ahead of Sprint 17. Sibling of 16-8.
   16-11-fix-card-deletion-cloud-resurrection: drafted # 2026-07-09: signed-in card delete resurrects on the next useCloudSync (pull-to-refresh/cold-open). Two uncoordinated sync engines: useCloudSync is deletion-blind — full-fetch + mergeCards re-adds cloud-only cards (cloud-sync.ts:337-340,:591) + re-upload (useCloudSync.ts:155-181); never drains syncPendingDeletions (deletion-tracker.ts:3) nor calls deleteCardFromCloud (cards.ts:75-87). Deletion-aware processPendingSync (useAutoSync -> sync-trigger.ts:128-142) loses the race. Purpose-built mergeWithDeletions (cloud-sync.ts:470-523) is DEAD CODE. Cousin of 16-8. Drafted by PM (John); Sprint 17. Needs refinement -> ready-for-dev.
   16-12-bound-ota-update-download-boot: drafted # 2026-07-10: Story 16.10 follow-up. checkForUpdateAsync was bounded in 16.10 (withTimeout 5s) but Updates.fetchUpdateAsync()/reloadAsync (app/_layout.tsx initializeApp) are NOT — a flaky connection that serves a manifest then stalls mid-bundle-download can still hang boot (online-only edge, out of 16.10's offline scope; pre-existing). Fix: wrap fetchUpdateAsync in withTimeout (core/utils/with-timeout.ts) behind a tunable budget that won't abort legit slow downloads; on timeout boot the current bundle. Drafted by Amelia (Dev); needs refinement -> ready-for-dev.
