+++ features/cards/index.ts
@@ -16,6 +16,12 @@
  * Exports card-related components, hooks, and repositories.
  */
 
+// Screens
+export { default as HomeScreen } from './screens/HomeScreen';
+export { default as CardDetailScreen } from './screens/CardDetailScreen';
+export { default as CardEditScreen } from './screens/CardEditScreen';
+export { default as BarcodeScreen } from './screens/BarcodeScreen';
+
 // Components
 export { CardList } from './components/CardList';
 export { CardTile } from './components/CardTile';
