+++ test/root-layout.offline-boot.test.tsx
@@ -13,7 +13,7 @@ import React from 'react';
 
 import { changeAppLanguage } from '@/shared/i18n';
 
-import RootLayout from '../_layout';
+import RootLayout from '@/app/_layout';
 
 const mockRouter = { replace: jest.fn(), back: jest.fn(), push: jest.fn() };
 const mockGetAllCards = jest.fn().mockResolvedValue([]);
